+++ src/ingest/github/files/prepare.rs
@@ -181,6 +181,7 @@ pub(super) async fn read_file_embed_docs(
                 content_type: "text",
                 title: Some(path.to_string()),
                 extra: Some(extra),
+                extractor_name: None,
             }
         })
         .collect();